Drop cluster.k8s.io, introduce machine.openshift.io
We are mirroring the API types under machine.openshift.io group. For the sake of transparency we are making this explicit in the import path. This allows both libraries upstream and openshift ot be consumed simultaneously by other tooling. We aim to consilidate back with upstream once there's a relase stable enough