e2e/autoscaler: consume MaxNodesTotalReached event
The previous behaviour was to scale out by waiting for the expected
scale up events to be generated on the node groups. After that had
completed we then waited for a further 60s to ensure that there were
no further scale up attempts as we had explicitly capped the cluster
size.

This commit changes the behaviour from waiting for 60s to explicitly
waiting for the MaxNodesTotalReached event to be generated. We use
that event to assert that the cluster will not grow beyond the
specified maximum.

This PR also helps to validate that we carry two upstream fixes that
were made with respect to calculating when the maximum size of the
cluster is reached:
- https://github.com/kubernetes/autoscaler/pull/1516
- https://github.com/kubernetes/autoscaler/pull/1771