e2e/autoscaler: create missing MachineSets as part of test setup
Rather than use the MachineSet's that may (or may not) come with the
installed platform we now create the number of MachineSet's we
actually need as part of test setup; on AWS we get three and on Azure
we currently get one. To handle these kind of differences this commit
changes the autoscaler test to create what it needs as part of test
setup.

The new setup requires at least one MachineSet with an accompanying
Node to be present. From that MachineSet we create the missing
MachineSet's - solely for autoscaler testing. Labels and the
ProviderSpec are propagated to the new MachineSet's from the first
MachineSet discovered on the cluster.