e2e/infra: ensure cluster size is restored after drain test
The test "drain node before removing machine resource" didn't wait for
the machine and node to be cleaned up on exit from the test. If the
autoscaler tests ran after this they would fail as there would be 1
more node in the cluster which would could be scaled-up. This only
occurs on Azure as it takes way longer for nodes (VMs, disks, et al)
to be deleted. And to compound matters (given the random ordering of
the tests) the e2e suite would only periodically fail; if the
autoscaler tests ran before node draining then they would complete
without error.