e2e/autoscaler: prevent scale up after workload has been deleted
We see occasional flakes in CI[1] where the test doesn't complete we
don't match the original machine count. This commit now waits for all
PODS in the workload to disappear (and the presence of these can
trigger further scale-ups) and then deletes the cluster-autoscaler
before scaling all the transient machinesets to zero. The intent is to
remove those resources that would allow a new machine/node to be
created.

The failure in CI looks like:

```
I0903 07:23:09.559723    7043 autoscaler.go:467] [1s remaining] Waiting for cluster to reach original machine count of 6; currently have 7

• Failure [1576.431 seconds]
[Feature:Machines] Autoscaler should
/tmp/tmp.TLbP279CTf/src/github.com/openshift/cluster-api-actuator-pkg/pkg/e2e/autoscaler/autoscaler.go:232
  scale up and down [It]
  /tmp/tmp.TLbP279CTf/src/github.com/openshift/cluster-api-actuator-pkg/pkg/e2e/autoscaler/autoscaler.go:233

  Timed out after 900.000s.
  Expected
      <int>: 7
  to equal
      <int>: 6
```

[1] https://openshift-gce-devel.appspot.com/build/origin-ci-test/logs/periodic-ci-openshift-cluster-autoscaler-operator-e2e-azure-master/272/