add a test for deletion annotations after scale down
This change adds a new test to the autoscaler suite for detecting
annotations that are left on Machines after a scale down to minimum
size. To accomplish this, several larger changes were added to the
suite.

One of the larger changes is that `Context` blocks have been added
around the 2 major tests cases. These blocks represent the different
ClusterAutoscalers that are used with each testing group. The first test
group, namely the "scale up and down" suite, has been left untouched.
The second test group, with "scale to/from zero", has been changed to
make the ClusterAutoscaler management part of the Before/After clauses
for each test.

This test is related to bugzilla case 1820410. reference:
https://bugzilla.redhat.com/show_bug.cgi?id=1820410