remove existingMachines and existingNodes variables
The Autoscaler "scale up and down" test used these variables to record
information but these usages do not have direct bearing on the test.
This change is being proposed to make the code easier to understand and
maintain. These checks do not exercise the tested behavior and appear to
be looking at the basic cluster size, the values are not used for
testing the desired behavior.