Wrap API errors in framework
By wrapping these errors, we preserve the underlying API error types 
which allows the errors to be read by the apimachinery errors package. 
This allows tests to determine the cause of an error and act 
appropriately. Eg you can tell if a resource does not exist vs a 
connection timeout error.