change autoscaler tests to use allocatable ram
This change switches the way the autoscaler tests calculate the usable
ram on worker nodes for the purpose of building resource requirements
for workload jobs. Prior to this commit, the tests have been using the
memory capacity of a node to determine the resource requirement. This
behavior can lead to failures in cases where the amount of actual
allocatable ram differs significantly from the capacity. This change
simply changes the calculation function to use the allocatable ram.