refactor autoscaler event watcher
This change moves the common functionality of the event watcher into the
global Describe scope for the autoscaler tests. It is being moved to
make it more widely available without duplicating code. As more tests
are added this event watcher proves more useful, it should be common to
all tests for the autoscaler.