rewrite scale up and down test for autoscaler
This change replaces the old test with a new version. This is being done
to move away from the event based testing of the previous version
towards an observation based test. What this means is that instead of
watching for scaling events the test will now observe the sizes of
MachineSet resources to determine scaling activity.