Add reconcile termination handler daemonSet validations
This https://github.com/openshift/machine-api-operator/pull/535 introduced support to manage a damonSet which runs termination handler for spot intances.
As an event handler is not passed to the damonSet informer changes to the resource won't trigger a reconcile.
This PR https://github.com/openshift/machine-api-operator/pull/648 fixes that by passing the event handler to the daemonSet namespaced informer.
This PR cover this e2e.