Merge spot tests to a single test
This means that the test suite only needs to spin up a single machineset 
for spot instances.

This is an easy alternative to having a `BeforeAll` which will hopefully 
come in Ginkgo V2, once it does, we can revert this and change the 
`BeforeEach` to a `BeforeAll`.

My hope is that by reducing the number of spot instances we need to spin 
up, this will improve the reliability of this test on Azure which seems 
to have some issues bringing up spot instances.