update autoscaler max nodes test
this change makes it so that the max nodes test will observe only the
nodes that are ready and schedulable in the cluster when determining
if the maximum has been reached. the autoscaler will only count ready
and schedulable nodes when performing its calculations, when this is
combined with tests that might not clean up the cluster properly, or
may leave nodes in a non schedulable state, then the test need to take
into account only the ready nodes when calculating the maximum size.