make autoscaler job names unique
this change makes the job names reflect the test they are running and
also make them unique to each test for the autoscaler. although these
workloads should not conflict with each other, they are all created in
the same namespace with the same name, this change will make them more
unique.