improve balance similar nodes tests
The balance similar nodes test has been quite flaky recently, this
change implements a number of fixes to improve the stability of this
test. The test has been updated to do the following:

* ensure the machinesets match on the labels they will apply to nodes
* add a wait for the machinesets to take the autoscaler annotations
* add a wait for the nodes to get the hugepages-1Gi and hugepages-2Mi
  resources
* add a wait for the nodes to have the same resource capacity keys
* refactor wait for scale out step to give more information

These changes are being added to address issues where the cluster
autoscaler is getting node records that appear different due to the
timing of when the hugepages resources are added.