add a sleep to balance similar nodes test
this change adds a 60 second sleep to the balance similar nodes test so
that the new nodes have proper time to acquire their capacity resources,
and so that the autoscaler has a few scan intervals to update its records.
through experimentation, it appears that 3-4 intervals is sufficient time
for the autoscaler to become aware of the changes. ideally this will be
remediated in the future, but for now this change is being proposed to
help the stability of this test.