MIXEDARCH-280: test scale from/to 0 with workloads using the kubernetes.io/arch label
When the workloads define a strong predicate on the kubernetes.io/arch label, the autoscaler should be able to foresee the possible architecture of the scaled node group even when it starts from 0 replicas currently and scale the proper one according to the requested architecture.
In the previous versions of OCP, any default labels (kubernetes.io/arch, kubernetes.io/os, ...) propagated from the MachineSet object to the NodeTemplate were overwritten in the ClusterAutoscaler's CAPI code with defaults and the autoscaling didn't work correctly for non-amd64 cases.

As of MIXEDARCH-129, we fixed the autoscaler and started changing the code of some machine-API providers (AWS, Azure today) to fill a capacity annotation by inspecting the instance type of the machine set's template and propagating the label of the future nodes being controlled to the node template in the autoscaler.

This case tests the scale from/to 0 with a workload defining a nodeSelectorRequirement for the architecture of the nodes, trying to bind a non-amd64 machine set.

Refers https://issues.redhat.com/browse/OCPBUGS-12745 and https://issues.redhat.com/browse/MIXEDARCH-129